Board

+++ src/main/java/Trendithon/SpinOff/domain/board/dto/BoardDto.java
@@ -12,7 +12,7 @@ public class BoardDto {
     private Long boardId;
     private String boardTitle; // 제목
     private String boardContext; // 내용
-    private Integer boardLike; // 좋아요
+    private Long boardLike; // 좋아요
     private List<String> imageUrl; // 이미지 URL
     private Long writer_id; // 작성자
 }
